backend/requirements.txt
fastapi
uvicorn
livekit-agents
livekit-plugins-openai
livekit-plugins-deepgram
livekit-plugins-cartesia
livekit-plugins-silero
livekit-plugins-google
livekit-plugins-noise-cancellation
python-dotenv
requests
aiohttp
mysql-connector-python
supabase
postgrest
